Import initial work for version 2
This change imports the ongoing initial work for version 2 of the
framework.  Version 2 will prove all currently-admitted lemmas.  To
achieve this, substantial work has been done to make the proof system
suitable for use in proving abstract assertions, such as those seen in
the derived rules like IIOI.

This is a work-in-progress.  At this point, neither component has been
updated to take advantage of the revised proof system and do not build.