Split components into multiple files
This change splits the functional definitions and specifications of
components into multiple individual files, allowing for parallel
compilation.
- The component file (i.e. perfect_link.v) exports the functional
  definition of the component and any lowered specifications from its
  sub-components.
- The separate specification files (i.e. perfect_link_1.v) export
  top-level properties of the component specification.
- The specification file (i.e. perfect_link_spec.v) exports all of the
  top-level properties of the component specifications.

Note that this change removes the incomplete proofs of PL_2, PL_3, and
their respective lemmas. They will be added as they are completed. The
goal from this point onward is to return to the system of each commit
producing a complete and error-free build.